organize Opsgenie provider classes
